+++ lotus/admin/__init__.py
@@ -1,8 +1,11 @@
+from .album import AlbumAdmin, AlbumItemAdmin
 from .article import ArticleAdmin
 from .category import CategoryAdmin
 
 
 __all__ = [
+    "AlbumAdmin",
+    "AlbumItemAdmin",
     "ArticleAdmin",
     "CategoryAdmin",
 ]