+++ lotus/factories/__init__.py
@@ -1,3 +1,4 @@
+from .album import AlbumFactory, AlbumItemFactory
 from .article import ArticleFactory, multilingual_article
 from .author import AuthorFactory
 from .category import CategoryFactory, multilingual_category
@@ -5,6 +6,8 @@ from .tag import TagFactory, TagNameBuilder
 
 
 __all__ = [
+    "AlbumFactory",
+    "AlbumItemFactory",
     "ArticleFactory",
     "AuthorFactory",
     "CategoryFactory",