+++ lotus/models/__init__.py
@@ -1,10 +1,13 @@
-from .author import Author
+from .album import Album, AlbumItem
 from .article import Article
+from .author import Author
 from .category import Category
 
 
 __all__ = [
-    "Author",
+    "Album",
+    "AlbumItem",
     "Article",
+    "Author",
     "Category",
 ]